Removes branch restrictions.

+++ deploy-ghpages.sh
@@ -11,15 +11,21 @@ gem install jekyll
 jekyll build --destination out
 echo "${DELIM}Finished build${DELIM}"
 
+# Branch
+MASTER="master"
+echo "${DELIM}Comparing ${MASTER} to ${TRAVIS_BRANCH}"
+
 # Deploy if on master branch
-echo "${DELIM}Starting deployment${DELIM}"
-cd out
-git init
-git config user.name "Travis-CI"
-git config user.email "travis@nodemeatspace.com"
-cp ../CNAME ./CNAME
-cp ../countryiso.js ./countryiso.js
-git add .
-git commit -m "Deployed to Github Pages"
-git push --force --quiet "https://${GH_TOKEN}@${GH_REF}" master:gh-pages > /dev/null 2>&1
-echo "${DELIM}Finished deployment${DELIM}"
+if [ "${MASTER}" = "${TRAVIS_BRANCH}" ]; then
+	echo "${DELIM}Starting deployment${DELIM}"
+	cd out
+	git init
+	git config user.name "Travis-CI"
+	git config user.email "travis@nodemeatspace.com"
+	cp ../CNAME ./CNAME
+	cp ../countryiso.js ./countryiso.js
+	git add .
+	git commit -m "Deployed to Github Pages"
+	git push --force --quiet "https://${GH_TOKEN}@${GH_REF}" master:gh-pages > /dev/null 2>&1
+	echo "${DELIM}Finished deployment${DELIM}"
+fi
